requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
fire==0.7.0
fonttools==4.54.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
imgaug==0.4.0
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
magic-filter==1.0.12
matplotlib==3.9.2
multidict==6.1.0
networkx==3.4.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.2
pillow==10.4.0
propcache==0.2.0
protobuf==5.28.2
pyclipper==1.3.0.post5
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
RapidFuzz==3.10.0
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
termcolor==2.5.0
tifffile==2024.9.20
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.2
